requirements.txt
bert-for-tf2==0.14.9
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
fonttools==4.45.0
gensim==4.3.2
idna==3.4
joblib==1.3.2
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.2
nltk==3.8.1
numpy==1.26.2
packaging==23.2
params-flow==0.8.2
Pillow==10.1.0
py-params==0.10.2
pymystem3==0.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-docx==1.1.0
regex==2023.10.3
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
summa==1.2.0
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
